Add TarEntry.ExtractToFileAsync tests
