+++ src/libraries/System.Formats.Tar/tests/System.Formats.Tar.Tests.csproj
@@ -10,6 +10,8 @@
     <Compile Include="$(CommonTestPath)System\IO\TempDirectory.cs" Link="Common\System\IO\TempDirectory.cs" />
     <Compile Include="CompressedTar.Async.Tests.cs" />
     <Compile Include="CompressedTar.Tests.cs" />
+    <Compile Include="TarEntry\TarEntry.ExtractToFile.Tests.cs" />
+    <Compile Include="TarEntry\TarEntry.ExtractToFileAsync.Tests.cs" />
     <Compile Include="TarFile\TarFile.ExtractToDirectoryAsync.File.Tests.Windows.cs" />
     <Compile Include="TarFile\TarFile.ExtractToDirectoryAsync.Stream.Tests.cs" />
     <Compile Include="TarFile\TarFile.ExtractToDirectoryAsync.File.Tests.cs" />
