+++ src/libraries/System.Formats.Tar/tests/TarEntry/TarEntryUstar.Tests.cs
@@ -89,78 +89,5 @@ namespace System.Formats.Tar.Tests
             SetFifo(fifo);
             VerifyFifo(fifo);
         }
-
-        [Fact]
-        public void Constructor_Name_FullPath_DestinationDirectory_Mismatch_Throws()
-        {
-            using TempDirectory root = new TempDirectory();
-
-            string fullPath = Path.Join(Path.GetPathRoot(root.Path), "dir", "file.txt");
-
-            UstarTarEntry entry = new UstarTarEntry(TarEntryType.RegularFile, fullPath);
-
-            entry.DataStream = new MemoryStream();
-            entry.DataStream.Write(new byte[] { 0x1 });
-            entry.DataStream.Seek(0, SeekOrigin.Begin);
-
-            Assert.Throws<IOException>(() => entry.ExtractToFile(root.Path, overwrite: false));
-
-            Assert.False(File.Exists(fullPath));
-        }
-
-        [Fact]
-        public void Constructor_Name_FullPath_DestinationDirectory_Match_AdditionalSubdirectory_Throws()
-        {
-            using TempDirectory root = new TempDirectory();
-
-            string fullPath = Path.Join(root.Path, "dir", "file.txt");
-
-            UstarTarEntry entry = new UstarTarEntry(TarEntryType.RegularFile, fullPath);
-
-            entry.DataStream = new MemoryStream();
-            entry.DataStream.Write(new byte[] { 0x1 });
-            entry.DataStream.Seek(0, SeekOrigin.Begin);
-
-            Assert.Throws<IOException>(() => entry.ExtractToFile(root.Path, overwrite: false));
-
-            Assert.False(File.Exists(fullPath));
-        }
-
-        [Fact]
-        public void Constructor_Name_FullPath_DestinationDirectory_Match()
-        {
-            using TempDirectory root = new TempDirectory();
-
-            string fullPath = Path.Join(root.Path, "file.txt");
-
-            UstarTarEntry entry = new UstarTarEntry(TarEntryType.RegularFile, fullPath);
-
-            entry.DataStream = new MemoryStream();
-            entry.DataStream.Write(new byte[] { 0x1 });
-            entry.DataStream.Seek(0, SeekOrigin.Begin);
-
-            entry.ExtractToFile(fullPath, overwrite: false);
-
-            Assert.True(File.Exists(fullPath));
-        }
-
-        [Theory]
-        [InlineData(TarEntryType.SymbolicLink)]
-        [InlineData(TarEntryType.HardLink)]
-        public void ExtractToFile_Link_Throws(TarEntryType entryType)
-        {
-            using TempDirectory root = new TempDirectory();
-            string fileName = "mylink";
-            string fullPath = Path.Join(root.Path, fileName);
-
-            string linkTarget = PlatformDetection.IsWindows ? @"C:\Windows\system32\notepad.exe" : "/usr/bin/nano";
-
-            UstarTarEntry entry = new UstarTarEntry(entryType, fileName);
-            entry.LinkName = linkTarget;
-
-            Assert.Throws<InvalidOperationException>(() => entry.ExtractToFile(fileName, overwrite: false));
-
-            Assert.Equal(0, Directory.GetFileSystemEntries(root.Path).Count());
-        }
     }
 }
